RC4 - update 1
Hello everyone, I really need some help on this project for school. I am not the greatest at coding but I really would love some help to get it done, it is due on Sunday, November 7th.. SO I really don't have too much time to get it done, but am committed in getting it done!